fix(ec2): `CfnInit` cannot be used with custom constructs (#11167)
Some users want to use `CfnInit` more flexibly, not with one of
our supported avenues of `Instance` or `AutoScalingGroup` (a prime
use case is AutoScalingGroups with LaunchTemplates, which we have no
L2 for yet).

To do that, they need to be able to call `CfnInit.attach()`.

This change makes that method public so users are able to call it.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
